Add integration with AWS S3 for image hosting Add ability to delete items from inventory
