+++ .env.local
@@ -1,1 +1,3 @@
 MONGODB_URI=mongodb+srv://rsgerrish:Jolene@cluster0.hltg7rw.mongodb.net/gmmg_shop?retryWrites=true&w=majority
+AWS_ACCESS_KEY_ID=AKIA53CHHLFWMCZOFVKP
+AWS_SECRET_ACCESS_KEY=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
