+++ package.json
@@ -18,6 +18,7 @@
     "@types/node": "18.14.5",
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
+    "aws-sdk": "^2.1333.0",
     "eslint": "8.35.0",
     "eslint-config-next": "13.2.3",
     "mongodb": "^5.1.0",
